WIP: History new text 'coming soon'. Change colour of the button of the side panel

+++ src/views/identities/containers/identities.js
@@ -1,14 +1,14 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { compose } from 'redux';
-import { Redirect, Switch } from 'react-router-dom';
 import { Map as ImmutableMap } from 'immutable';
 import {
+  Box,
   Button,
   Widget,
 } from 'base_components';
+import * as BOX_CONSTANTS from 'constants/box';
 import { withIdentities } from 'hocs';
-import * as ROUTES from 'constants/routes';
 import List from '../components/list';
 
 
@@ -34,27 +34,63 @@ class Identities extends Component {
   };
 
   state = {
-    redirectToWizard: false,
+    askForDelete: false,
+  };
+
+  /**
+   * Show or not the box with the confirmation for deleting the identities.
+   */
+  toggleAskConfirmation = () => {
+    this.setState(prevState => ({ askForDelete: !prevState.askForDelete }));
   };
 
   /**
    * Call the action to delete all the identities in the app state
-   * and in the storage, and redirec to to the wizard to create an identity
+   * and in the storage, and redirect to to the wizard to create an identity
    */
   deleteAllIdentities = () => {
-    this.props.handleDeleteAllIdentities()
-      .then(() => this.setState({ redirectToWizard: true }));
+    // TODO: An error is displayed in the console. Check it
+    this.toggleAskConfirmation();
+    this.state.askForDelete && this.props.handleDeleteAllIdentities();
   };
+
+  /**
+   * The content to show in the confirmation box to delete all de identities.
+   * There is a warning text and the button to delete them.
+   *
+   * @returns {Object} React element with the content to show
+   * @private
+   */
+  _getConfirmationContent() {
+    return (
+      <div className="i3-ww-identities__delete-confirmation">
+        <div className="i3-ww-identities__delete-confirmation-text">
+          You are about to delete all the identities of in this application
+          and device. This action can not be undone. If you continue, you
+          accept that all data stored will deleted.
+        </div>
+        <div className="i3-ww-identities__delete-confirmation-button">
+          <Button
+            type="primary"
+            htmlType="button"
+            onClick={this.deleteAllIdentities}>
+            I understand it: Delete them!
+          </Button>
+        </div>
+      </div>
+    );
+  }
 
   render() {
     const headerButtons = (
       <Button
         type="primary"
         htmlType="button"
-        onClick={this.deleteAllIdentities}>
+        onClick={this.toggleAskConfirmation}>
         Delete all
       </Button>
     );
+    const confirmationContent = this._getConfirmationContent();
 
     return (
       <div className="i3-ww-identities">
@@ -69,10 +105,15 @@ class Identities extends Component {
             togglePinned={this.togglePinned}
             list={this.props.identities} />
         </Widget>
-        {
-          this.state.redirectToWizard
-          && <Redirect to={ROUTES.CREATE_IDENTITY.MAIN} />
-        }
+        <div>
+          <Box
+            type={BOX_CONSTANTS.TYPE.SIDE_PANEL}
+            side={BOX_CONSTANTS.SIDE.RIGHT}
+            onClose={this.toggleAskConfirmation}
+            content={confirmationContent}
+            title="Delete all the identities"
+            show={this.state.askForDelete} />
+        </div>
       </div>
     );
   }
